+++ inc/woocommerce/shipping/single-product-shipping.php
@@ -38,6 +38,7 @@ if ( ! function_exists('single_product_manufacturer_shipping_and_info') ) :
       'Q' => [ //QUOIZEL
         'shipping' => 'Ships in 2 to 3 Business Days',
         'warranty' => 'https://www.quoizel.com/warranty/',
+        'coupon-code-quoizel' => 'Get 15% Off This Item With Code "QUOIZELALL15" at Checkout!'
         ],
   
       '0' => [ //SAVOY