+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable theme changes should be added to this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [ released ] - Scheduled for release 10.31.2020
+
+## [ 1.27.0 ]
+
+## Removed
+
+- Maxim outdoor lighting sale from sale page
+
 ## [ released ] - 10.29.2020
 
 ## [ 1.26.0 ]